pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fewshot-hc-nas"
version = "0.1.0"
description = "Few-shot hardware-conditioned NAS prototype for unknown MCU and edge devices."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.1.0",
  "pandas>=2.2.0",
  "pyyaml>=6.0",
  "matplotlib>=3.9.0",
  "torch>=2.1.0",
  "pyserial>=3.5",
]

[project.optional-dependencies]
paper = ["pyarrow>=18.0.0"]
dev = ["pytest>=8.0.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
